refactored per john request
